Cpp:: and Crystal:: helper functionality
Let's be honest: One can't live without utility classes.  Without
duplicating much code, that is.  So let's accept the fact that we can't
without, and make the best of it: These modules contain a ton of
`struct`s, which have are specialized in functionality and language.
Both languages have many structs of the same name, and much of their
methods are called the same (or similar).  Hopefully, this choice
reduces cognitive load, as you only have to learn 1.5 sets of methods,
instead of 2.0.

Bindgen::TypeHelper is shared functionality between those structures,
across languages.  I'm sure we'll find a better name later on for this.