Processors: Make use of the graph
The new architecture enables us to do many things not possible before.
My favorites include:

* Automatic container instantiation (Those were annoying to type)
* Sanity checks, to tell you your config is broken, and how
* Graph dumping for better insight into what will be generated
* Modularity in general!  Even method filtering is done here.

Now where's light there's shadow, and here it's that the user now has
to configure the pipeline(s) manually.  I hope the new possiblities
outweigh this a bit.  Also, the TEMPLATE.yml comes with a preconfigured
pipeline.

Fun fact: This alone makes it possible for wrapper developers to add
custom functionality to bindgen.  Which I think is really cool. However
this wasn't even my intention, so you could say it's an accidential
feature.  And a darn useful one at that.