Processor::FunctionClass: Wrap C APIs with classes!
This is one of those features I always wanted in bindgen.  Far too
often have I dismissed an idea because it involved using some C lib,
and I didn't want to mess with bare bindings.  While you still have
to write configuration for this to work, in my honest opinion, this is
still so much better than fiddling with C bindings, leaking pointers,
and then writing "just a quick wrapper" to avoid it.  Those days be
gone!