Version 0.6: Portable insanity
I think the recent additions are worthy of a version bump.  There's
still a lot to be done, that's for sure.  But we're about 60% there :)

This release offers the (just recently added) find paths feature, which
lets you find .. paths of your dependencies.  With enough creative path
globs, and calls to shell snippets, one can build somewhat portable
bindings configurations somewhat easily :)

This feature is also used internally for the find_clang.cr script to
find an appropriate clang++ binary.  That alone is enough to find all
other paths from it - And thus, makes bindgen portable-ish!

I'm happy to announce that next to ArchLinux, bindgen now also works
on Ubuntu 16.04 machines.  It should work anywhere it finds a clang++
binary, as pathing assumptions are now gone (In the clang part).  But
I can't promise it to just work everywhere ;)

That's it!