clang/find_clang.cr: Auto-detect if linking libclang using .a or .so
ArchLinux has changed their mind in the meantime.  Didn't feel like
messing with the environment variable, instead, it's supposed to
auto-detect now if libclang.a or libclang.so exist.

If both static and dynamic versions are available it'll prefer the
static library, which is hopefully more stable when the user decides
to update their system.

The user can still use BINDGEN_DYNAMIC=1/0 as before, but simply not
setting it will trigger the auto-detection.

Also, the pattern-matching style "case" statement of Crystal is some
good stuff!