Fix non-pointer constructors for struct like types.
The bindgen generated code assumes that a class construction
expression returns a pointer. This is not true for value-type classes
(structs), hence a subsequent dereferencing is invalid C++.

This change modifies the generated construction code (only for types
of kind "Struct") such that the value of a construction expression can
always be referenced. This is achieved by a simple C++ template
wrapper "bg_deref".

It also modifies the generated Crystal code. The `@unwrap` and
`to_unsafe` methods now return the plain C++-struct. The only place
where a pointer is actually used is the `self` argument. Because
`to_unsafe` will not work in this case, the generator now explicitly
emits `pointerof(@unwrap)` for the self argument.